chore: document the code and change the name handlers for repositories
